Add a struct for passing around configuration
This adds a `Config` struct that can be used to pass around the
configuration. A `load_configuration_from_environment` function is being
added to populate the struct.

This has two benefits: 1) it simplifies `main` as it no longer needs to
repeat itself for each configuration setting and 2) it should make the
code easier to test because (future) functions can accept a `Config`
that doesn't have to be populated based on environment variables.