Copy the files
This copies any files added to the source folder into the destination
folder. Doing so requires updating `create_parent_path_if_missing` to
accept a reference to the path rather than its value. (And since I had
to update the function's signature I decided to go ahead and move it to
the correctly alphabetized location.)